+++ Framework/ConvivaSDK.xcframework/ios-arm64_armv7/ConvivaSDK.framework/PrivateHeaders/CISClient.h
@@ -45,5 +45,8 @@
 
 -(void)detachPlayer:(int)sessionKey monitorMetrics:(BOOL) monitorMetrics;
 
+- (int) getGlobalSessionId;
+- (int) getipv4SessionId;
+- (int) getipv6SessionId;
 
 @end