+++ Framework/ConvivaSDK.xcframework/ios-arm64_i386_x86_64-simulator/ConvivaSDK.framework/PrivateHeaders/CISStreamAnalyticsImpl+Private.h
@@ -50,6 +50,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void) reportDroppedFramesTotalCount:(int64_t)droppedFramesTotal;
 
+- (void) reportAudioLanguage:(NSString *)audioLanguage;
+
+- (void) reportClosedCaptionLanguage:(NSString *)closedCaptionLanguage;
+
+- (void) reportSubtitleLanguage:(NSString *)subtitleLanguage;
+
 - (void)reportAirPlayEvents;
 
 @end